feature/principal-master Added Address support and refactored code.

+++ src/main/java/com/alsharif/operations/shipprincipal/entity/AddressDetails.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -14,7 +15,7 @@ import java.time.LocalDateTime;
 public class AddressDetails {
     @Id
     @Column(name = "ADDRESS_POID", nullable = false)
-    private Long addressPoid;
+    private BigDecimal addressPoid;
 
     @Column(name = "ADDRESS_MASTER_POID", nullable = false)
     private Long addressMasterPoid;
